Added Connector test
