+++ src/main/java/io/cdap/plugin/sfmc/connector/MarketingConnectorConfig.java
@@ -16,30 +16,18 @@
 
 package io.cdap.plugin.sfmc.connector;
 
-import com.exacttarget.fuelsdk.ETApiObject;
-import com.exacttarget.fuelsdk.ETDataExtensionRow;
 import com.exacttarget.fuelsdk.ETSdkException;
-import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Strings;
 import io.cdap.cdap.api.annotation.Description;
 import io.cdap.cdap.api.annotation.Macro;
 import io.cdap.cdap.api.annotation.Name;
-import io.cdap.cdap.api.data.format.StructuredRecord;
 import io.cdap.cdap.api.data.schema.Schema;
 import io.cdap.cdap.api.plugin.PluginConfig;
 import io.cdap.cdap.etl.api.FailureCollector;
 import io.cdap.plugin.sfmc.source.MarketingCloudClient;
 import io.cdap.plugin.sfmc.source.MarketingCloudInputFormat;
 import io.cdap.plugin.sfmc.source.util.MarketingCloudConstants;
-import io.cdap.plugin.sfmc.source.util.MarketingCloudObjectInfo;
 import io.cdap.plugin.sfmc.source.util.SourceObject;
 import io.cdap.plugin.sfmc.source.util.Util;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import java.lang.reflect.Method;
-import java.util.Date;
-import java.util.List;
-import java.util.concurrent.TimeUnit;
 
 /**
  * Salesforce Marketing Cloud Connector Config
@@ -153,115 +141,9 @@ public class MarketingConnectorConfig extends PluginConfig {
               this.getAuthEndpoint(),
               this.getSoapEndpoint());
       return MarketingCloudInputFormat.getTableMetaData(sourceObject,
-              null,  client).getSchema();
+              null, client).getSchema();
     } catch (ETSdkException e) {
       throw new RuntimeException(e);
-    }
-  }
-  private static final Logger LOG = LoggerFactory.getLogger(MarketingConnectorConfig.class);
-
-  private Object getFieldValue(ETApiObject row, String fieldName) {
-    try {
-      Method method = row.getClass().getMethod(createGetterName(fieldName));
-      return method.invoke(row);
-    } catch (Exception e) {
-      LOG.error(String.format("Error while fetching %s.%s value", row.getClass().getSimpleName(), fieldName), e);
-      return null;
-    }
-  }
-  private String createGetterName(String name) {
-    StringBuilder sb = new StringBuilder("get");
-    sb.append(name.substring(0, 1).toUpperCase());
-    sb.append(name.substring(1));
-    return sb.toString();
-  }
-
-  @VisibleForTesting
-  String convertToStringValue(Object fieldValue) {
-    return String.valueOf(fieldValue);
-  }
-
-  @VisibleForTesting
-  Double convertToDoubleValue(Object fieldValue) {
-    if (fieldValue instanceof String && Strings.isNullOrEmpty(String.valueOf(fieldValue))) {
-      return null;
-    }
-
-    return Double.parseDouble(String.valueOf(fieldValue));
-  }
-
-  @VisibleForTesting
-  Integer convertToIntegerValue(Object fieldValue) {
-    if (fieldValue instanceof String && Strings.isNullOrEmpty(String.valueOf(fieldValue))) {
-      return null;
-    }
-
-    return Integer.parseInt(String.valueOf(fieldValue));
-  }
-
-  @VisibleForTesting
-  Boolean convertToBooleanValue(Object fieldValue) {
-    if (fieldValue instanceof String && Strings.isNullOrEmpty(String.valueOf(fieldValue))) {
-      return null;
-    }
-
-    return Boolean.parseBoolean(String.valueOf(fieldValue));
-  }
-  @VisibleForTesting
-  Object convertToValue(String fieldName, Schema fieldSchema, Object fieldValue) {
-    Schema.Type fieldType = fieldSchema.getType();
-    Schema.LogicalType logicalType = fieldSchema.getLogicalType();
-    if (fieldSchema.getLogicalType() != null) {
-      return transformLogicalType(fieldName, logicalType, fieldValue);
-    }
-    switch (fieldType) {
-      case STRING:
-        return convertToStringValue(fieldValue);
-      case DOUBLE:
-        return convertToDoubleValue(fieldValue);
-      case INT:
-        return convertToIntegerValue(fieldValue);
-      case BOOLEAN:
-        return convertToBooleanValue(fieldValue);
-      case UNION:
-        if (fieldSchema.isNullable()) {
-          return convertToValue(fieldName, fieldSchema.getNonNullable(), fieldValue);
-        }
-        throw new IllegalStateException(
-                String.format("Field '%s' is of unexpected type '%s'. Declared 'complex UNION' types: %s",
-                        fieldName, fieldValue.getClass().getSimpleName(), fieldSchema.getUnionSchemas()));
-      default:
-        throw new IllegalStateException(
-                String.format("Record type '%s' is not supported for field '%s'", fieldType.name(), fieldName));
-    }
-  }
-  public void convertRecord(MarketingCloudObjectInfo sfObjectMetaData, StructuredRecord.Builder recordBuilder, ETApiObject row) throws ETSdkException {
-
-    List<Schema.Field> tableFields = sfObjectMetaData.getSchema().getFields();
-    for (Schema.Field field : tableFields) {
-      String fieldName = field.getName();
-      Object rawFieldValue = null;
-      if (row instanceof ETDataExtensionRow) {
-        String apiFieldName = sfObjectMetaData.lookupFieldsMap(fieldName);
-        rawFieldValue = ((ETDataExtensionRow) row).getColumn(apiFieldName);
-      } else {
-        rawFieldValue = getFieldValue(row, fieldName);
-      }
-      Object fieldValue = convertToValue(fieldName, field.getSchema(), rawFieldValue);
-
-      recordBuilder.set(fieldName, fieldValue);
-    }
-  }
-  private Object transformLogicalType(String fieldName, Schema.LogicalType logicalType, Object value) {
-    switch (logicalType) {
-      case TIMESTAMP_MICROS:
-        if (value instanceof Date) {
-          return TimeUnit.MILLISECONDS.toMicros((((Date) value).getTime()));
-        }
-        return null;
-      default:
-        throw new IllegalArgumentException(
-                String.format("Field '%s' is of unsupported type '%s'", fieldName, logicalType.getToken()));
     }
   }
 }
